requirements.txt
gradio
pandas
faiss-cpu
sentence-transformers
google-generativeai
scikit-learn
